Use uint (not long) compare

+++ src/DotNet/Emit/MethodBodyReaderBase.cs
@@ -497,7 +497,7 @@ namespace dnlib.DotNet.Emit {
 		/// <param name="index">A parameter index</param>
 		/// <returns>A <see cref="Parameter"/> or <c>null</c> if <paramref name="index"/> is invalid</returns>
 		protected Parameter GetParameter(int index) {
-			if ((uint)index >= parameters.Count)
+			if ((uint)index >= (uint)parameters.Count)
 				return null;
 			return parameters[index];
 		}
@@ -508,7 +508,7 @@ namespace dnlib.DotNet.Emit {
 		/// <param name="index">A local index</param>
 		/// <returns>A <see cref="Local"/> or <c>null</c> if <paramref name="index"/> is invalid</returns>
 		protected Local GetLocal(int index) {
-			if ((uint)index >= locals.Count)
+			if ((uint)index >= (uint)locals.Count)
 				return null;
 			return locals[index];
 		}
